Address review: move From impls to domain modules per cloud-identity-wallet pattern
