+++ src/database/error.rs
@@ -23,3 +23,24 @@ impl From<sea_orm::DbErr> for RepositoryError {
         RepositoryError::Generic(err.to_string())
     }
 }
+
+impl From<RepositoryError> for crate::web::errors::ApiError {
+    fn from(err: RepositoryError) -> Self {
+        tracing::error!(error = %err, "repository error");
+        let description = err.to_string();
+        let (status, error_code) = match &err {
+            RepositoryError::DuplicateEntry => {
+                (axum::http::StatusCode::CONFLICT, std::borrow::Cow::Borrowed("duplicate_entry"))
+            }
+            _ => (
+                axum::http::StatusCode::INTERNAL_SERVER_ERROR,
+                std::borrow::Cow::Borrowed("internal_error"),
+            ),
+        };
+        crate::web::errors::ApiError {
+            status,
+            error: error_code,
+            error_description: Some(description),
+        }
+    }
+}
