+++ src/web/errors.rs
@@ -8,10 +8,7 @@ use axum::{
 };
 use serde::{Deserialize, Serialize};
 
-use crate::database::error::RepositoryError;
-use crate::web::auth::errors::AuthenticationError;
 use crate::web::handlers::issue_credential::CredentialError;
-use crate::web::handlers::status_list::error::StatusListError;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ApiError {
@@ -68,47 +65,6 @@ impl IntoResponse for ApiError {
         };
 
         (self.status, Json(body)).into_response()
-    }
-}
-
-impl From<StatusListError> for ApiError {
-    fn from(err: StatusListError) -> Self {
-        Self {
-            status: err.get_status(),
-            error: err.get_error_code(),
-            error_description: Some(err.get_error_message()),
-        }
-    }
-}
-
-impl From<AuthenticationError> for ApiError {
-    fn from(err: AuthenticationError) -> Self {
-        Self {
-            status: err.get_status(),
-            error: err.get_error_code(),
-            error_description: Some(err.get_error_message()),
-        }
-    }
-}
-
-impl From<RepositoryError> for ApiError {
-    fn from(err: RepositoryError) -> Self {
-        tracing::error!(error = %err, "repository error");
-        let description = err.to_string();
-        let (status, error_code) = match &err {
-            RepositoryError::DuplicateEntry => {
-                (StatusCode::CONFLICT, Cow::Borrowed("duplicate_entry"))
-            }
-            _ => (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Cow::Borrowed("internal_error"),
-            ),
-        };
-        ApiError {
-            status,
-            error: error_code,
-            error_description: Some(description),
-        }
     }
 }
