+++ src/web/handlers/status_list/error.rs
@@ -3,6 +3,8 @@ use std::borrow::Cow;
 use axum::http::StatusCode;
 use thiserror::Error;
 
+use crate::web::errors::ApiError;
+
 #[derive(Debug, Clone, PartialEq, Eq, Error)]
 pub enum StatusListError {
     #[error("Invalid list ID string: {0}")]
@@ -90,6 +92,16 @@ impl StatusListError {
 
     pub(crate) fn get_error_message(&self) -> String {
         self.to_string()
+    }
+}
+
+impl From<StatusListError> for ApiError {
+    fn from(err: StatusListError) -> Self {
+        Self {
+            status: err.get_status(),
+            error: err.get_error_code(),
+            error_description: Some(err.get_error_message()),
+        }
     }
 }
